pyproject.toml
[project]
name = "generative-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-mistralai>=1.1.2",
    "langchain-openai>=1.1.14",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
]
